api/claim_verification/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
babel==2.18.0
bitarray==3.8.0
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
colbert-ai==0.2.22
coloredlogs==15.0.1
confection==0.1.5
courlan==1.3.2
cymem==2.0.13
Cython==3.2.4
datasets==4.5.0
dateparser==1.2.2
dill==0.4.0
exceptiongroup==1.3.1
faiss-cpu==1.13.0
filelock==3.19.1
Flask==3.1.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.2
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jusText==3.0.2
lightgbm==4.6.0
lxml==6.0.2
lxml_html_clean==0.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.2.1
ninja==1.13.0
nmslib==2.1.2
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.19.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
pyarrow==21.0.0
pybind11==3.0.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyjnius==1.7.0
pyserini==0.22.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
thinc==8.3.9
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
trafilatura==2.0.0
transformers==4.57.6
triton==3.4.0
typer==0.23.2
typer-slim==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.6
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
